# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HRM Laravel Api",
        "version": "10.0"
    },
    "paths": {
...
        "/api/subscriber": {
            "get": {
                "tags": [
                    "Subscriber"
                ],
                "summary": "Get All subscribers.Permission required = subscriber.list",
                "description": "This endpoint retrieves information about something.",
                "operationId": "1de97ff2fcf6d9f26c8f614bf1176134",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Subscriber"
                ],
                "summary": "Create a new subscriber.Permission required = subscriber.store",
                "description": "This endpoint creates a new subscriber.",
                "operationId": "cf7efcfbae40c18a9096ae20727a789b",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "tenant_id": {
                                        "description": "The id of the tenant => required",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "subscription_plan_id": {
                                        "description": "The id of the subscription_plan => required",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "start_date": {
                                        "description": "The starting date of plan=> required and its format should be yyyy-mm-dd",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-03-25"
                                    },
                                    "end_date": {
                                        "description": "The ending date of plan=> required and its format should be yyyy-mm-dd",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-03-25"
                                    },
                                    "amount": {
                                        "description": "The amount of plan=> required",
                                        "type": "string",
                                        "example": "100 PKR"
                                    },
                                    "status": {
                                        "description": "The status of the subscriber =>required",
                                        "type": "number",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "subscriber created successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            }
        },
        "/api/subscriber/inactive-subscribers": {
            "get": {
                "tags": [
                    "Subscriber"
                ],
                "summary": "Get All subscribers.Permission required = subscriber.list",
                "description": "This endpoint retrieves information about something.",
                "operationId": "6bdd10ff2890b8c4e804326704339970",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/subscriber/create": {
            "get": {
                "tags": [
                    "Subscriber"
                ],
                "summary": "Get All companies that are active.Permission required = subscriber.create",
                "description": "This endpoint retrieves information about something.",
                "operationId": "b95f694801449599c298d4ee38e704fa",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/subscriber/{id}": {
            "get": {
                "tags": [
                    "Subscriber"
                ],
                "summary": "GET The subscriber.Permission required = subscriber.edit",
                "description": "This endpoint Gives a specific  subscriber.",
                "operationId": "35af6f135c36ecdd1b7a9d95d9118300",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the subscriber ",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Subscriber"
                ],
                "summary": "Delete The subscriber.Permission required = subscriber.delete",
                "description": "This endpoint delete subscriber.",
                "operationId": "5c3c3bac17db4422cdf5a260be8ebeb0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the subscriber to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Subscriber"
                ],
                "summary": "Update the subscriber.Permission required = subscriber.update",
                "description": "This endpoint updates a subscriber.",
                "operationId": "c45a1856bd08e689a258f51551720dfa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the subscriber to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "tenant_id": {
                                        "description": "The id of the tenant => required",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "subscription_plan_id": {
                                        "description": "The id of the subscription_plan => required",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "start_date": {
                                        "description": "The starting date of plan=> required and its format should be yyyy-mm-dd",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-03-25"
                                    },
                                    "end_date": {
                                        "description": "The ending date of plan=> required and its format should be yyyy-mm-dd",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-03-25"
                                    },
                                    "amount": {
                                        "description": "The amount of plan=> required",
                                        "type": "string",
                                        "example": "100 PKR"
                                    },
                                    "status": {
                                        "description": "The status of the subscriber =>required",
                                        "type": "number",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "subscriber created successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}